+++ app/Http/Controllers/DrawController.php
@@ -33,43 +33,33 @@ class DrawController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
         $request->validate([
             'draw_title' => 'required',
+        ], [
+            'draw_title.required' => 'Il titolo è obbligatorio',
         ]);
 
-        $draw = new Draw();
-        $draw->user_id = Auth::id();
-        $draw->draw_title = $request->draw_title;
-        $draw->draw_url = $request->draw_url;
+        try {
+            $draw = new Draw();
+            $draw->user_id = Auth::id();
+            $draw->draw_title = $request->draw_title;
+            $draw->draw_url = $request->draw_url;
 
-        if (array_key_exists('draw_img', $request->all())) {
-            $url_file = Storage::put('/draw_imgs', $request['draw_img']);
-            $draw->draw_img = $url_file;
+            if (array_key_exists('draw_img', $request->all())) {
+                $url_file = Storage::put('/draw_imgs', $request['draw_img']);
+                $draw->draw_img = $url_file;
+            }
+
+            $draw->save();
+
+            return back()->with('message', 'Contenuto aggiunto');
+        } catch (Exception $e) {
+            return back()->with('message', 'Qualcosa è andato storto');
         }
-
-        $draw->save();
-
-        return back()->with('message', 'Contenuto aggiunto');
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(Draw $draw)
-    {
-        //
     }
 
     /**
@@ -77,7 +67,9 @@ class DrawController extends Controller
      */
     public function edit(Draw $draw)
     {
-        return Inertia::render('Draws/Edit', compact('draw'));
+        if ($draw->user_id == Auth::id()) {
+            return Inertia::render('Draws/Edit', compact('draw'));
+        }
     }
 
     /**
@@ -87,19 +79,30 @@ class DrawController extends Controller
     {
         $request->validate([
             'draw_title' => 'required',
+        ], [
+            'draw_title.required' => 'Il titolo è obbligatorio',
         ]);
-        if ($request->hasfile('draw_img')) {
-            Storage::delete($draw->draw_img);
 
-            $url_file = Storage::put('/draw_imgs', $request->file('draw_img'));
-            $draw->draw_img = $url_file;
+        try {
+
+            if ($draw->user_id == Auth::id()) {
+
+                if ($request->hasfile('draw_img')) {
+                    Storage::delete($draw->draw_img);
+
+                    $url_file = Storage::put('/draw_imgs', $request->file('draw_img'));
+                    $draw->draw_img = $url_file;
+                }
+                $draw->update([
+                    'draw_title' => $request->draw_title,
+                    'draw_url' => $request->draw_url
+                ]);
+
+                return to_route('draws.index')->with('message', 'Contenuto aggiornato');
+            }
+        } catch (Exception $e) {
+            return back()->with('message', 'Qualcosa è andato storto');
         }
-        $draw->update([
-            'draw_title' => $request->draw_title,
-            'draw_url' => $request->draw_url
-        ]);
-
-        return to_route('draws.index')->with('message', 'Contenuto aggiornato');
     }
 
     /**
@@ -107,12 +110,20 @@ class DrawController extends Controller
      */
     public function destroy(Draw $draw)
     {
-        if ($draw->draw_img) {
-            Storage::delete($draw->draw_img);
+        try {
+
+            if ($draw->user_id == Auth::id()) {
+
+                if ($draw->draw_img) {
+                    Storage::delete($draw->draw_img);
+                }
+
+                $draw->delete();
+
+                return back()->with('message', 'Contenuto eliminato');
+            }
+        } catch (Exception $e) {
+            return back()->with('message', 'Qualcosa è andato storto');
         }
-
-        $draw->delete();
-
-        return back()->with('message', 'Contenuto eliminato');
     }
 }
